UI/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
outetts
uroman
torch
torchaudio
transformers
numpy
cloudinary
yarngpt